[review] リストアできるように

+++ src/restoreAuth.ts
@@ -3,27 +3,30 @@ import { Storage } from '@google-cloud/storage';
 import { auth } from 'firebase-tools';
 import { readFile, unlink, writeFile } from 'fs/promises';
 
-export const backupAuth = async ({
+export const restoreAuth = async ({
   region,
+  encriptedFilePath,
   projectId = process.env.GCLOUD_PROJECT,
   bucketName = `${process.env.GCLOUD_PROJECT}-authentication-backups`,
 }: {
   region: string;
+  encriptedFilePath: string;
   projectId?: string;
   bucketName?: string;
 }): Promise<void> => {
   const plaintextFileName = `firebase-authentication-backup.csv`;
   const tmpPlaintextFileName = `/tmp/${plaintextFileName}`;
   const tmpCiphertextFileName = `/tmp/${plaintextFileName}.encripted`;
-  const gcsDestination = `${new Date().toISOString()}/${plaintextFileName}.encripted`;
 
-  // ローカルに取得
-  await auth.export(tmpPlaintextFileName, { project: projectId });
+  // GCS から ローカルに取得
+  const gcsClient = new Storage();
+  const bucket = gcsClient.bucket(bucketName);
+  await bucket.file(encriptedFilePath).download({ destination: tmpCiphertextFileName });
 
   // ファイル読み込み
-  const plaintext = await readFile(tmpPlaintextFileName);
+  const ciphertext = await readFile(tmpCiphertextFileName);
 
-  // 暗号化
+  // 復号化
   const kmsClient = new KeyManagementServiceClient();
   const keyName = kmsClient.cryptoKeyPath(
     projectId,
@@ -31,13 +34,18 @@ export const backupAuth = async ({
     'firebase-authentication-keyring',
     'firebase-authentication-backup-key'
   );
-  const [result] = await kmsClient.encrypt({ name: keyName, plaintext });
-  await writeFile(tmpCiphertextFileName, result.ciphertext as string);
+  const [result] = await kmsClient.decrypt({
+    name: keyName,
+    ciphertext: ciphertext,
+  });
 
-  // GCS に保存
-  const gcsClient = new Storage();
-  const bucket = gcsClient.bucket(bucketName);
-  await bucket.upload(tmpCiphertextFileName, { destination: gcsDestination });
+  if (!result.plaintext) {
+    throw new Error('Decrypt Failed.');
+  }
+  await writeFile(tmpPlaintextFileName, result.plaintext.toString());
+
+  // Authの復元
+  await auth.upload(tmpPlaintextFileName, { project: projectId });
 
   // ローカルのファイルを削除
   await Promise.all([unlink(tmpPlaintextFileName), unlink(tmpCiphertextFileName)]);
